+++ src/main/java/com/ks/ny/entity/vo/SteelPipeExcelVO.java
@@ -2,6 +2,9 @@ package com.ks.ny.entity.vo;
 
 import com.alibaba.excel.annotation.ExcelProperty;
 
+import java.math.BigDecimal;
+
+
 public class SteelPipeExcelVO {
 
     @ExcelProperty("记录ID(不可修改)")
@@ -83,29 +86,29 @@ public class SteelPipeExcelVO {
     private String f0000118;
     //长/mm
     @ExcelProperty("长/mm")
-    private Double length;
+    private Integer length;
     //宽/mm
     @ExcelProperty("宽/mm")
-    private Double width;
+    private Integer width;
 
     //入库总数量/mm
     @ExcelProperty("入库总数量/mm")
-    private Double inboundLength;
+    private Integer inboundLength;
     //流程进行中的数量/mm
     @ExcelProperty("流程进行中的数量/mm")
-    private Double processLength;
+    private Integer processLength;
     //消耗的数量/mm
     @ExcelProperty("消耗的数量/mm")
-    private Double quantityConsumed;
+    private Integer quantityConsumed;
     //剩余库存数量/mm
     @ExcelProperty("剩余库存数量/mm")
-    private Double inventoryLength;
+    private Integer inventoryLength;
     //理论剩余重量/kg
     @ExcelProperty("理论剩余重量/kg")
-    private Double inventoryWeight;
+    private Integer inventoryWeight;
     //质保号库存数量mm
     @ExcelProperty("质保号库存数量mm")
-    private Double warrantyNumberLength;
+    private Integer warrantyNumberLength;
 
 
     public String getObjectId() {
@@ -315,67 +318,67 @@ public class SteelPipeExcelVO {
         this.f0000118 = f0000118;
     }
 
-    public Double getLength() {
+    public Integer getLength() {
         return length;
     }
 
-    public void setLength(Double length) {
+    public void setLength(Integer length) {
         this.length = length;
     }
 
-    public Double getWidth() {
+    public Integer getWidth() {
         return width;
     }
 
-    public void setWidth(Double width) {
+    public void setWidth(Integer width) {
         this.width = width;
     }
 
-    public Double getInboundLength() {
+    public Integer getInboundLength() {
         return inboundLength;
     }
 
-    public void setInboundLength(Double inboundLength) {
+    public void setInboundLength(Integer inboundLength) {
         this.inboundLength = inboundLength;
     }
 
-    public Double getProcessLength() {
+    public Integer getProcessLength() {
         return processLength;
     }
 
-    public void setProcessLength(Double processLength) {
+    public void setProcessLength(Integer processLength) {
         this.processLength = processLength;
     }
 
-    public Double getQuantityConsumed() {
+    public Integer getQuantityConsumed() {
         return quantityConsumed;
     }
 
-    public void setQuantityConsumed(Double quantityConsumed) {
+    public void setQuantityConsumed(Integer quantityConsumed) {
         this.quantityConsumed = quantityConsumed;
     }
 
-    public Double getInventoryLength() {
+    public Integer getInventoryLength() {
         return inventoryLength;
     }
 
-    public void setInventoryLength(Double inventoryLength) {
+    public void setInventoryLength(Integer inventoryLength) {
         this.inventoryLength = inventoryLength;
     }
 
-    public Double getInventoryWeight() {
+    public Integer getInventoryWeight() {
         return inventoryWeight;
     }
 
-    public void setInventoryWeight(Double inventoryWeight) {
+    public void setInventoryWeight(Integer inventoryWeight) {
         this.inventoryWeight = inventoryWeight;
     }
 
-    public Double getWarrantyNumberLength() {
+    public Integer getWarrantyNumberLength() {
         return warrantyNumberLength;
     }
 
-    public void setWarrantyNumberLength(Double warrantyNumberLength) {
+    public void setWarrantyNumberLength(Integer warrantyNumberLength) {
         this.warrantyNumberLength = warrantyNumberLength;
     }
 }
